pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quantdl"
version = "0.1.1"
description = "Financial data fetcher for alpha research with S3 backend and local caching"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "QuantDL Team" }]
keywords = ["finance", "quantitative", "data", "polars", "alpha"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
    "Typing :: Typed",
]
dependencies = [
    "polars>=1.0.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "ruff>=0.14.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "moto[s3]>=5.0.0",
    "boto3>=1.34.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.urls]
Homepage = "https://github.com/quantdl/quantdl-api"
Documentation = "https://github.com/quantdl/quantdl-api#readme"
Repository = "https://github.com/quantdl/quantdl-api"

[tool.hatch.build.targets.wheel]
packages = ["src/quantdl"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]
extend-exclude = ["*.ipynb"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"examples/**/*.py" = ["E402"]  # load_dotenv() before imports is intentional

[tool.ruff.lint.isort]
known-first-party = ["quantdl"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = []

[[tool.mypy.overrides]]
module = ["moto", "boto3", "botocore", "scipy", "scipy.stats", "numpy"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/quantdl"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "python-dotenv>=1.2.1",
]
